requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.1.40
browserbase==1.4.0
bubus==1.5.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.29.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
instructor==1.10.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.3
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.9
langgraph==0.2.76
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.3.45
litellm==1.75.7
loguru==0.7.3
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==0.14.1
MarkupSafe==3.0.2
mcp==1.13.0
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.3
openai==1.99.9
openpyxl==3.1.5
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
patchright==1.52.5
playwright==1.54.0
pluggy==1.6.0
portalocker==2.10.1
portia-sdk-python==0.6.2
posthog==6.5.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdf==6.0.0
pyperclip==1.9.0
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
screeninfo==0.8.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
testcontainers==4.12.0
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uv==0.8.11
uvicorn==0.35.0
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
